'requires_ansible' does not need to be semver
Import was checking if the requires_ansible field was
valid semver. But the format used by ansible for versioning
is not valid semver.

Instead it is based on pythons builting
packagaging.specifiers.Specifier and SpecifierSet

This changes replaces the validation of 'requires_ansible'
with one based on SpecifierSet.

Issue: AAH-981