Show more info when Legacy role galaxy info has unknown field
Currently during legacy role import, if galaxy_info has an unknown
field, we see the following message:

  unknown field in galaxy_info

This is not very helpful, and does not tell us how to fix the issue.

This change adds the cause of the issue to the exception message, so we
know which field to fix.

No-Issue

Signed-off-by: Mark Goddard <mark@stackhpc.com>